requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
concurrent_log==1.0.1
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.5
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.56
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.27.0
orjson==3.10.3
packaging==23.2
pluggy==1.5.0
portalocker==2.8.2
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
pytest==8.2.0
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-llm==0.7.1
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
yarl==1.9.4
